+++ opencv_interop/CMakeLists.txt
@@ -6,13 +6,13 @@ set( OpenCV_DEPENDENCY
     "find_dependency( OpenCV )" PARENT_SCOPE
 )
 
-file( GLOB_RECURSE HEADERS src/*.h src/*.inl )
-file( GLOB_RECURSE SOURCES src/*.cpp )
+file( GLOB_RECURSE HEADERS *.h *.inl )
+file( GLOB_RECURSE SOURCES *.cpp )
 
 add_library( cgt_opencv_interop SHARED ${HEADERS} ${SOURCES} )
 target_include_directories(cgt_core
     PUBLIC
-		$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>
+		$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../..>
         $<INSTALL_INTERFACE:include>
 )
 target_link_libraries( cgt_opencv_interop cgt_core ${OpenCV_LIBS} )
@@ -25,6 +25,6 @@ install( TARGETS cgt_opencv_interop
 	INCLUDES DESTINATION include
 )
 
-install( DIRECTORY src/ DESTINATION include/libcgt/opencv_interop
+install( DIRECTORY . DESTINATION include/libcgt/opencv_interop
     FILES_MATCHING PATTERN "*.h" PATTERN "*.inl"
 )
